+++ doc/releaseNotes.md
@@ -207,3 +207,5 @@ This file describes the main feature changes for each readout.exe released versi
   - added options:
     - dumpDataInline: if set, each packet raw content is printed (hex dump style).
     - fileReadVerbose: if set, more information is printed when reading/decoding file. By default, file size/chunking printouts are now off.
+- Updated configuration parameters:
+  - added readout.memoryPoolStatsEnabled